Spit homepage components into files, fix some type errors
