+++ src/components/pages-component/home-page/index.ts
@@ -1,4 +1,9 @@
+export * from './form-add-new-coin';
 export * from './add-token-form';
+export * from './form-add-transaction';
+export * from './board';
+export * from './button-center-modal';
+export * from './button-left-side-modal';
 export * from './modal-center';
 export * from './modal-left-side';
-export * from './other';
+export * from './token';
