+++ src/components/pages-component/home-page/token.tsx
@@ -1,0 +1,20 @@
+export function Token({
+  token,
+  tokenDelete,
+}: {
+  token: any;
+  tokenDelete: (token: any) => { void: any };
+}) {
+  const handleDelete = () => {
+    tokenDelete(token);
+  };
+
+  return (
+    <div>
+      {token.symbol}: {token.price}|
+      <button type="button" onClick={handleDelete}>
+        Delete
+      </button>
+    </div>
+  );
+}
